+++ src/views/components/pages/index/template.tsx
@@ -5,24 +5,20 @@ import MessageEvaluateRequestData from '~/src/server/interfaces/message/evaluate
 import MessageEvaluateResponseData from '~/src/server/interfaces/message/evaluate/POST/Response';
 import MessageTranslateRequestData from '~/src/server/interfaces/message/translate/POST/Request';
 import MessageTranslateResponseData from '~/src/server/interfaces/message/translate/POST/Response';
-import { FaLongArrowAltDown, FaCheck } from 'react-icons/fa';
-import { HiSelector } from 'react-icons/hi';
+import { FaLongArrowAltDown } from 'react-icons/fa';
 import { GenerateButton } from '../../domain/GenerateButton/template';
 import MessageForm from '../../domain/MessageForm';
 import { BaseTemplate } from '../../template/BaseTemplate';
-import { Listbox } from '@headlessui/react';
-import clsx from 'clsx';
+import { Mode, ModeSelect } from '../../template/ModeSelect';
 
 const CASUAL_VALUE = 70;
 
-export type Mode = 'evaluate' | 'translate';
-
-const ModeList: Mode[] = ['evaluate', 'translate'];
+const modeList: Mode[] = ['evaluate', 'translate'];
 
 export const RootPage: NextPage = () => {
   const [evaluatedMessageValue, setEvaluatedMessageValue] = useState<string>('');
   const [messageFormValue, setMessageFormValue] = useState<string>('');
-  const [currentMode, setCurrentMode] = useState<Mode>(ModeList[0]);
+  const [currentMode, setCurrentMode] = useState<Mode>(modeList[0]);
 
   const {
     responseData: evaluateValue,
@@ -57,30 +53,9 @@ export const RootPage: NextPage = () => {
 
   return (
     <BaseTemplate>
-      <Listbox value={currentMode} onChange={setCurrentMode}>
-        <Listbox.Button className='flex bg-red-400 text-white w-fit px-[32px] justify-center items-center'>
-          <div className='font-bold text-[24px]'>モードを変更する</div>
-          <HiSelector className='w-[32px] h-[32px]' />
-        </Listbox.Button>
-        <Listbox.Options className='absolute w-fit bg-white m-[4px] p-[8px] rounded-[8px]'>
-          {ModeList.map((mode, idx) => (
-            <Listbox.Option
-              key={idx}
-              value={mode}
-              className={({ active }) =>
-                clsx('cursor-pointer px-[8px]', active ? 'bg-red-100 text-red-900' : 'text-gray-900')
-              }
-            >
-              {({ selected }) => (
-                <div className='flex items-center justify-center'>
-                  {selected && <FaCheck />}
-                  <div className={selected ? 'font-medium' : 'font-normal'}>{mode}</div>
-                </div>
-              )}
-            </Listbox.Option>
-          ))}
-        </Listbox.Options>
-      </Listbox>
+      <ModeSelect currentMode={currentMode} setCurrentMode={setCurrentMode} modeList={modeList}>
+        モードを変更する
+      </ModeSelect>
       {currentMode === 'evaluate' && (
         <div className='space-y-[8px] flex items-center justify-center flex-col'>
           <div>{evaluatedMessageValue}</div>
